requirements.txt
PyQt5>=5.15.0
pandas>=1.3.0
chardet>=4.0.0
requests>=2.25.0
openai-whisper>=20230918
matplotlib>=3.5.0
numpy>=1.21.0
openpyxl>=3.0.0
wordcloud>=1.8.0
reportlab>=3.6.0
pyaudio>=0.2.11